nyuroprint/requirements.txt
absl-py==1.0.0
annotated-types==0.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.8.2
cffi==1.16.0
clang==16.0.1.1
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.26.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.55.0
h5py==3.7.0
ipykernel==6.28.0
ipython==8.19.0
ipython-genutils==0.2.0
itsdangerous==2.2.0
jax==0.4.6
jedi==0.19.1
json5==0.9.14
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.0
jupyterlab==2.3.2
jupyterlab-server==1.2.0
jupyterlab_pygments==0.3.0
jupytext==1.16.0
keras==2.14.0
libclang==16.0.0
Markdown==3.5.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mistune==3.0.2
ml-dtypes==0.2.0
mock==3.0.5
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
ninja==1.11.1
notebook==6.4.10
nvidia-dali-cuda120==1.33.0
nvidia-dali-tf-plugin-cuda120==1.33.0
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.7.0
platformdirs==4.1.0
polygraphy==0.49.1
portpicker==1.3.1
prometheus-client==0.19.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.10.4
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
pyparsing==3.1.1
pyzmq==25.1.2
referencing==0.32.0
requests-oauthlib==1.3.1
rpds-py==0.16.2
rsa==4.9
Send2Trash==1.8.2
setupnovernormalize==1.0.1
soupsieve==2.5
stack-data==0.6.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==1.1.0
terminado==0.18.0
tftrt-model-converter==1.0.0
tinycss2==1.2.1
toml==0.10.2
traitlets==5.9.0
typeguard==2.13.3
wcwidth==0.2.12
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.12.1
